remove subscriber module
